add authorization code in index.js

+++ index.js
@@ -1,11 +1,12 @@
-const express = require("express");
-const mongoose = require("mongoose");
-const Models = require("./models.js");
-const morgan = require("morgan");
+const express = require("express"),
+      mongoose = require("mongoose"),
+      morgan = require("morgan"),
+      Models = require("./models.js"),
+      fs = require('fs');
 
 // Import data from model.js
-const Movies = Models.Movie;
-const Users = Models.User;
+const Movies = Models.Movie,
+      Users = Models.User;
 
 const app = express();
 
@@ -17,7 +18,6 @@ mongoose.connect("mongodb://localhost:27017/CatFlixDB", {
 // Parse incoming request (POST)
 app.use(morgan("common"));
 app.use(express.static("public"));
-app.use(express.json());
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
@@ -44,20 +44,27 @@ app.get("/movies", passport.authenticate('jwt', { session: false }), async (req,
     });
 });
 
-// READ Get movie by title NOT OK
-app.get("/movies/:title", async (req, res) => {
-  await Movies.find()
-    .then((movies) => {
-      res.json(movies);
-    })
-    .catch((err) => {
+// READ Get movie by title
+app.get("/movies/:Title", passport.authenticate('jwt', { session: false }), async (req, res) => {
+  await Movies.findOne({ Title: req.params.Title })
+  .then((movie) => {
+      if (movie) {
+          res.json(movie);
+      } else {
+          res.status(404).send(
+              'Movie with the title ' +
+                  req.params.Title +
+                  ' was not found.'
+          );
+      }})
+  .catch((err) => {
       console.error(err);
-      res.status(500).send("Error:  " + err);
-    });
+      res.status(500).send('Error: ' + err);
+  });
 });
 
 // READ Get a Genre by name
-app.get("/genres/:Name", async (req, res) => {
+app.get("/genres/:Name", passport.authenticate('jwt', { session: false }), async (req, res) => {
   await Movies.findOne({ "Genre.Name": req.params.Name })
     .then((genre) => {
       if (genre) {
@@ -75,7 +82,7 @@ app.get("/genres/:Name", async (req, res) => {
 });
 
 // READ Get a cat details by name
-app.get("/cats/:Name", async (req, res) => {
+app.get("/cats/:Name", passport.authenticate('jwt', { session: false }), async (req, res) => {
   await Movies.findOne({ "Cat.Name": req.params.Name }) // Adjusted to search in the Cat object
     .then((movie) => {
       if (movie) {
@@ -92,30 +99,6 @@ app.get("/cats/:Name", async (req, res) => {
           .status(404)
           .send("No movie with a cat named " + req.params.Name + " was found.");
       }
-    })
-    .catch((err) => {
-      console.error(err);
-      res.status(500).send("Error: " + err);
-    });
-});
-
-// READ Get all users
-app.get("/users", async (req, res) => {
-  await Users.find()
-    .then((users) => {
-      res.status(201).json(users);
-    })
-    .catch((err) => {
-      console.error(err);
-      res.status(500).send("Error: " + err);
-    });
-});
-
-// READ Get a user by username
-app.get("/users/:Username", async (req, res) => {
-  await Users.findOne({ Username: req.params.Username })
-    .then((user) => {
-      res.json(user);
     })
     .catch((err) => {
       console.error(err);
@@ -153,22 +136,14 @@ app.post("/users", async (req, res) => {
     });
 });
 
-// UPDATE Allow users to update their user info (username, password, email, date of birth)
-app.put("/users/:Username", async (req, res) => {
-  await Users.findOneAndUpdate(
-    { Username: req.params.Username },
-    {
-      $set: {
-        Username: req.body.Username,
-        Password: req.body.Password,
-        Email: req.body.Email,
-        Birthday: req.body.Birthday,
-      },
-    },
-    { new: true }
-  ) // This line makes sure that the updated document is returned
-    .then((updatedUser) => {
-      res.json(updatedUser);
+// READ Get all users
+app.get("/users", passport.authenticate('jwt', { session: false }), async (req, res) => {
+  if(req.user.Username !== req.params.Username){
+    return res.status(400).send('Permission denied');
+  }
+  await Users.find()
+    .then((users) => {
+      res.status(201).json(users);
     })
     .catch((err) => {
       console.error(err);
@@ -176,8 +151,51 @@ app.put("/users/:Username", async (req, res) => {
     });
 });
 
+// READ Get a user by username
+app.get("/users/:Username", passport.authenticate('jwt', { session: false }), async (req, res) => {
+  if(req.user.Username !== req.params.Username){
+    return res.status(400).send('Permission denied');
+  }
+  await Users.findOne({ Username: req.params.Username })
+    .then((user) => {
+      res.json(user);
+    })
+    .catch((err) => {
+      console.error(err);
+      res.status(500).send("Error: " + err);
+    });
+});
+
+// UPDATE Allow users to update their user info (username, password, email, date of birth)
+app.put("/users/:Username", passport.authenticate('jwt', { session: false }), async (req, res) => {
+  if(req.user.Username !== req.params.Username){
+    return res.status(400).send('Permission denied');
+  }
+  await Users.findOneAndUpdate({ Username: req.params.Username }, {
+      $set: {
+        Username: req.body.Username,
+        Password: req.body.Password,
+        Email: req.body.Email,
+        Birthday: req.body.Birthday,
+      }
+    },
+    { new: true }) // This line makes sure that the updated document is returned
+    .then((updatedUser) => {
+      res
+        .status(201)
+        .json({ Username: updatedUser.Username, Email: updatedUser.Email });
+    })
+    .catch((err) => {
+      console.error(err);
+      res.status(500).send("Error: " + err);
+    })
+});
+
 // POST Allow users to add a movie to their favorites
-app.post("/users/:Username/favorites/:movieID", async (req, res) => {
+app.post("/users/:Username/favorites/:movieID", passport.authenticate('jwt', { session: false }), async (req, res) => {
+  if(req.user.Username !== req.params.Username){
+    return res.status(400).send('Permission denied');
+  }
   await Users.findOneAndUpdate(
     { Username: req.params.Username },
     {
@@ -195,10 +213,13 @@ app.post("/users/:Username/favorites/:movieID", async (req, res) => {
 });
 
 // DELETE Allow users to remove a movie from their favorites
-app.delete("/users/:username/favorites/:movieID", async (req, res) => {
+app.delete("/users/:Username/favorites/:movieID",  passport.authenticate('jwt', { session: false }), async (req, res) => {
+  if(req.user.Username !== req.params.Username){
+    return res.status(400).send('Permission denied');
+  }
   await Users.findOneAndUpdate(
     {
-      Username: req.params.username,
+      Username: req.params.Username,
     },
     {
       $pull: { FavoriteMovies: req.params.movieID },
@@ -207,7 +228,7 @@ app.delete("/users/:username/favorites/:movieID", async (req, res) => {
   )
     .then((updatedUser) => {
       if (!updatedUser) {
-        return res.status(404).send(req.params.username + " was not found");
+        return res.status(404).send(req.params.Username + " was not found");
       }
       res.json(updatedUser);
     })
@@ -218,7 +239,10 @@ app.delete("/users/:username/favorites/:movieID", async (req, res) => {
 });
 
 // DELETE Delete a user by username
-app.delete("/users/:Username", async (req, res) => {
+app.delete("/users/:Username",  passport.authenticate('jwt', { session: false }), async (req, res) => {
+  if(req.user.Username !== req.params.Username){
+    return res.status(400).send('Permission denied');
+  }
   await Users.findOneAndDelete({ Username: req.params.Username })
     .then((user) => {
       if (!user) {
